# api/docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "APIs SNM",
    "description": "Documentazione APIs: SNM - Social Network for Music"
  },
  "host": "localhost:3100",
  "basePath": "/",
  "tags": [
    {
      "name": "esplora",
      "description": "Endpoints per la sezione generale esplora"
    },
    {
      "name": "auth",
      "description": "Endpoints per la gestione delle autenticazioni"
    },
    {
      "name": "info",
      "description": "Endpoints per la gestione di informazioni personali"
    },
    {
      "name": "playlist",
      "description": "Endpoints per la gestione delle proprie playlist personali/importate"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/{username}": {
      "get": {
        "tags": [
          "auth"
        ],
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/registrazione": {
      "post": {
        "tags": [
          "auth"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/delete/{username}": {
      "delete": {
        "tags": [
          "info"
        ],
        "description": "",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/change/": {
      "put": {
        "tags": [
          "info"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/changeArtistsFav/": {
      "put": {
        "tags": [
          "info"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/artistsFav/": {
      "get": {
        "tags": [
          "info"
        ],
        "description": "",
        "parameters": [
          {
            "name": "iduser",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/changeGenresFav/": {
      "put": {
        "tags": [
          "info"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/genresFav/": {
      "get": {
        "tags": [
          "info"
        ],
        "description": "",
        "parameters": [
          {
            "name": "iduser",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/playlistPersonali/add": {
      "post": {
        "tags": [
          "playlist"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/playlistPersonali/view": {
      "get": {
        "tags": [
          "playlist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/playlistPersonali/viewOne": {
      "get": {
        "tags": [
          "playlist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "idplaylist",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/playlistPersonali/delete": {
      "delete": {
        "tags": [
          "playlist"
        ],
        "description": "",
        "parameters": [
          {
            "name": "idplaylist",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/playlistPersonali/addTrack": {
      "put": {
        "tags": [
          "playlist"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/playlistPersonali/deleteOne": {
      "put": {
        "tags": [
          "playlist"
        ],
        "description": "",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/esplora": {
      "get": {
        "tags": [
          "esplora"
        ],
        "description": "",
        "parameters": [
          {
            "name": "iduser",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/esplora/playlistPubblica": {
      "get": {
        "tags": [
          "esplora"
        ],
        "description": "",
        "parameters": [
          {
            "name": "idplaylist",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/esplora/playlistPubblica/by": {
      "get": {
        "tags": [
          "esplora"
        ],
        "description": "",
        "parameters": [
          {
            "name": "usersPub",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}